ref: precise exception handler into separate module
We can move the other lints later if we feel like it.
