+++ pyproject.toml
@@ -32,11 +32,11 @@ ipython = "^7.2"
 
 [tool.poetry.plugins]
 [tool.poetry.plugins."flake8.extension"]
-PIE781 = "flake8_pie:Flake8PieCheck"
-PIE783 = "flake8_pie:Flake8PieCheck"
-PIE784 = "flake8_pie:Flake8PieCheck"
-PIE785 = "flake8_pie:Flake8PieCheck"
-PIE786 = "flake8_pie:Flake8PieCheck"
+PIE781 = "flake8_pie:Flake8PieCheck781"
+PIE783 = "flake8_pie:Flake8PieCheck783"
+PIE784 = "flake8_pie:Flake8PieCheck784"
+PIE785 = "flake8_pie:Flake8PieCheck785"
+PIE786 = "flake8_pie:Flake8PieCheck786"
 
 [build-system]
 requires = ["poetry>=0.12"]
